@@ -1,35 +1,11 @@
-// Define versions in a single place
-ext {
-    // Sdk and tools
-    minSdkVersion = 14
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '26.0.2'
-
-    // App dependencies
-    supportLibraryVersion = '26.1.0'
-    guavaVersion = '22.0-android'
-    junitVersion = '4.12'
-    mockitoVersion = '2.8.47'
-    hamcrestVersion = '1.3'
-    runnerVersion = '1.0.1'
-    rulesVersion = '1.0.1'
-    espressoVersion = '3.0.1'
-    roomVersion = "1.0.0"
-
-    // Architecture Components dependencies
-    archLifecycleVersion = "1.0.0"
-}
-
 buildscript {
-    ext.kotlinVersion = '1.1.51'
-    ext.gradlePluginVersion = '3.0.0'
+    ext.kotlinVersion = '1.3.11'
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:$gradlePluginVersion"
+        classpath 'com.android.tools.build:gradle:3.3.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -39,11 +15,64 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 }
 
 task clean(type: Delete) {
     delete rootProject.buildDir
+}
+
+// Define versions in a single place
+ext {
+    // Sdk and tools
+    // Support library and architecture components support minSdk 14 and above.
+    minSdkVersion = 14
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+
+    // App dependencies
+    androidXVersion = '1.0.0'
+    androidXTestVersion = '1.1.1'
+    androidXAnnotations = '1.0.1'
+    androidXLegacySupport = '1.0.0'
+    materialVersion = '1.0.0'
+    cardVersion = '1.0.0'
+    recyclerViewVersion = '1.0.0'
+    appCompatVersion = '1.0.2'
+
+    guavaVersion = '27.0.1-android'
+    junitVersion = '4.12'
+    mockitoVersion = '2.8.9'
+    dexMakerVersion = '2.12.1'
+    hamcrestVersion = '1.3'
+    runnerVersion = '1.0.1'
+    rulesVersion = '1.0.1'
+    espressoVersion = '3.1.1'
+    roomVersion = '2.0.0'
+    archLifecycleVersion = '2.0.0'
+    archTestingVersion = '1.1.1'
+}
+
+/*
+ * Workaround for https://code.google.com/p/android/issues/detail?id=182715
+ *
+ * The Android Gradle plugin is creating DataBindingExportBuildInfoTasks for the instrumentation
+ * APK that generates from the app APKs layouts. This creates duplicate classes in the app and
+ * instrumentation APK which leads to
+ *      java.lang.IllegalAccessError: Class ref in pre-verified class resolved to unexpected implementation
+ * on older devices.
+ *
+ * The workaround is to get the DataBindingExportBuildInfoTasks tasks for the instrumentation APK
+ * and delete the files right after it creates them.
+ */
+subprojects {
+    tasks.withType(com.android.build.gradle.internal.tasks.databinding.DataBindingExportBuildInfoTask) { task ->
+        if (task.name.endsWith("AndroidTest")) {
+            task.finalizedBy(tasks.create("${task.name}Workaround") << {
+                task.output.deleteDir()
+            })
+        }
+    }
 }
