@@ -15,42 +15,46 @@
  */
 package com.example.android.architecture.blueprints.todoapp.tasks
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.InstrumentationRegistry.getTargetContext
-import android.support.test.espresso.Espresso.onView
-import android.support.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu
-import android.support.test.espresso.action.ViewActions.click
-import android.support.test.espresso.action.ViewActions.closeSoftKeyboard
-import android.support.test.espresso.action.ViewActions.replaceText
-import android.support.test.espresso.action.ViewActions.typeText
-import android.support.test.espresso.assertion.ViewAssertions.doesNotExist
-import android.support.test.espresso.assertion.ViewAssertions.matches
-import android.support.test.espresso.matcher.ViewMatchers.hasSibling
-import android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom
-import android.support.test.espresso.matcher.ViewMatchers.isChecked
-import android.support.test.espresso.matcher.ViewMatchers.isDescendantOfA
-import android.support.test.espresso.matcher.ViewMatchers.isDisplayed
-import android.support.test.espresso.matcher.ViewMatchers.withContentDescription
-import android.support.test.espresso.matcher.ViewMatchers.withId
-import android.support.test.espresso.matcher.ViewMatchers.withText
-import android.support.test.filters.LargeTest
-import android.support.test.filters.SdkSuppress
-import android.support.test.rule.ActivityTestRule
-import android.support.test.runner.AndroidJUnit4
 import android.view.View
 import android.widget.ListView
+import androidx.test.InstrumentationRegistry.getTargetContext
+import androidx.test.core.app.ApplicationProvider
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu
+import androidx.test.espresso.IdlingRegistry
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.action.ViewActions.closeSoftKeyboard
+import androidx.test.espresso.action.ViewActions.replaceText
+import androidx.test.espresso.action.ViewActions.typeText
+import androidx.test.espresso.assertion.ViewAssertions.doesNotExist
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.hasSibling
+import androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom
+import androidx.test.espresso.matcher.ViewMatchers.isChecked
+import androidx.test.espresso.matcher.ViewMatchers.isDescendantOfA
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.withContentDescription
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.filters.LargeTest
+import androidx.test.filters.SdkSuppress
+import androidx.test.rule.ActivityTestRule
+import androidx.test.runner.AndroidJUnit4
 import com.example.android.architecture.blueprints.todoapp.Injection
 import com.example.android.architecture.blueprints.todoapp.R
+import com.example.android.architecture.blueprints.todoapp.R.string
 import com.example.android.architecture.blueprints.todoapp.ViewModelFactory
 import com.example.android.architecture.blueprints.todoapp.currentActivity
 import com.example.android.architecture.blueprints.todoapp.getToolbarNavigationContentDescription
 import com.example.android.architecture.blueprints.todoapp.rotateOrientation
+import com.example.android.architecture.blueprints.todoapp.util.EspressoIdlingResource
 import com.google.common.base.Preconditions.checkArgument
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.TypeSafeMatcher
 import org.hamcrest.core.IsNot.not
+import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -59,7 +63,9 @@ import org.junit.runner.RunWith
 /**
  * Tests for the tasks screen, the main screen which contains a list of all tasks.
  */
-@RunWith(AndroidJUnit4::class) @LargeTest class TasksScreenTest {
+@RunWith(AndroidJUnit4::class)
+@LargeTest
+class TasksScreenTest {
 
     private val TITLE1 = "TITLE1"
     private val TITLE2 = "TITLE2"
@@ -73,12 +79,32 @@ import org.junit.runner.RunWith
      * blocks of Junit tests.
      */
     @Rule @JvmField var tasksActivityTestRule =
-            ActivityTestRule<TasksActivity>(TasksActivity::class.java)
+        ActivityTestRule<TasksActivity>(TasksActivity::class.java)
 
-    @Before fun resetState() {
+    @Before
+    fun resetState() {
         ViewModelFactory.destroyInstance()
-        Injection.provideTasksRepository(InstrumentationRegistry.getTargetContext())
-                .deleteAllTasks()
+        Injection.provideTasksRepository(ApplicationProvider.getApplicationContext())
+            .deleteAllTasks()
+    }
+
+    /**
+     * Prepare your test fixture for this test. In this case we register an IdlingResources with
+     * Espresso. IdlingResource resource is a great way to tell Espresso when your app is in an
+     * idle state. This helps Espresso to synchronize your test actions, which makes tests
+     * significantly more reliable.
+     */
+    @Before
+    fun registerIdlingResource() {
+        IdlingRegistry.getInstance().register(EspressoIdlingResource.countingIdlingResource)
+    }
+
+    /**
+     * Unregister your Idling Resource so it can be garbage collected and does not leak any memory.
+     */
+    @After
+    fun unregisterIdlingResource() {
+        IdlingRegistry.getInstance().unregister(EspressoIdlingResource.countingIdlingResource)
     }
 
     private val toolbarNavigationContentDescription: String
@@ -98,17 +124,18 @@ import org.junit.runner.RunWith
         checkArgument(itemText.isNotEmpty(), "itemText cannot be null or empty")
         return object : TypeSafeMatcher<View>() {
             override fun matchesSafely(item: View) = allOf(
-                    isDescendantOfA(isAssignableFrom(ListView::class.java)),
-                    withText(itemText)
+                isDescendantOfA(isAssignableFrom(ListView::class.java)),
+                withText(itemText)
             ).matches(item)
 
             override fun describeTo(description: Description) {
-                description.appendText("is isDescendantOfA LV with text " + itemText)
+                description.appendText("is isDescendantOfA LV with text $itemText")
             }
         }
     }
 
-    @Test fun clickAddTaskButton_opensAddTaskUi() {
+    @Test
+    fun clickAddTaskButton_opensAddTaskUi() {
         // Click on the add task button
         onView(withId(R.id.fab_add_task)).perform(click())
 
@@ -116,7 +143,8 @@ import org.junit.runner.RunWith
         onView(withId(R.id.add_task_title)).check(matches(isDisplayed()))
     }
 
-    @Test fun editTask() {
+    @Test
+    fun editTask() {
         // First add a task
         createTask(TITLE1, DESCRIPTION)
 
@@ -131,9 +159,11 @@ import org.junit.runner.RunWith
 
         // Edit task title and description
         onView(withId(R.id.add_task_title))
-                .perform(replaceText(editTaskTitle), closeSoftKeyboard()) // Type new task title
-        onView(withId(R.id.add_task_description)).perform(replaceText(editTaskDescription),
-                closeSoftKeyboard()) // Type new task description and close the keyboard
+            .perform(replaceText(editTaskTitle), closeSoftKeyboard()) // Type new task title
+        onView(withId(R.id.add_task_description)).perform(
+            replaceText(editTaskDescription),
+            closeSoftKeyboard()
+        ) // Type new task description and close the keyboard
 
         // Save the task
         onView(withId(R.id.fab_edit_task_done)).perform(click())
@@ -145,14 +175,16 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE1)).check(doesNotExist())
     }
 
-    @Test fun addTaskToTasksList() {
+    @Test
+    fun addTaskToTasksList() {
         createTask(TITLE1, DESCRIPTION)
 
         // Verify task is displayed on screen
         onView(withItemText(TITLE1)).check(matches(isDisplayed()))
     }
 
-    @Test fun markTaskAsComplete() {
+    @Test
+    fun markTaskAsComplete() {
         viewAllTasks()
 
         // Add active task
@@ -170,7 +202,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE1)).check(matches(isDisplayed()))
     }
 
-    @Test fun markTaskAsActive() {
+    @Test
+    fun markTaskAsActive() {
         viewAllTasks()
 
         // Add completed task
@@ -189,7 +222,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE1)).check(matches(not(isDisplayed())))
     }
 
-    @Test fun showAllTasks() {
+    @Test
+    fun showAllTasks() {
         // Add 2 active tasks
         createTask(TITLE1, DESCRIPTION)
         createTask(TITLE2, DESCRIPTION)
@@ -200,7 +234,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE2)).check(matches(isDisplayed()))
     }
 
-    @Test fun showActiveTasks() {
+    @Test
+    fun showActiveTasks() {
         // Add 2 active tasks
         createTask(TITLE1, DESCRIPTION)
         createTask(TITLE2, DESCRIPTION)
@@ -211,7 +246,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE2)).check(matches(isDisplayed()))
     }
 
-    @Test fun showCompletedTasks() {
+    @Test
+    fun showCompletedTasks() {
         // Add 2 completed tasks
         createTask(TITLE1, DESCRIPTION)
         clickCheckBoxForTask(TITLE1)
@@ -224,7 +260,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE2)).check(matches(isDisplayed()))
     }
 
-    @Test fun clearCompletedTasks() {
+    @Test
+    fun clearCompletedTasks() {
         viewAllTasks()
 
         // Add 2 complete tasks
@@ -235,14 +272,15 @@ import org.junit.runner.RunWith
 
         // Click clear completed in menu
         openActionBarOverflowOrOptionsMenu(getTargetContext())
-        onView(withText(R.string.menu_clear)).perform(click())
+        onView(withText(string.menu_clear)).perform(click())
 
         //Verify that completed tasks are not shown
         onView(withItemText(TITLE1)).check(matches(not(isDisplayed())))
         onView(withItemText(TITLE2)).check(matches(not(isDisplayed())))
     }
 
-    @Test fun createOneTask_deleteTask() {
+    @Test
+    fun createOneTask_deleteTask() {
         viewAllTasks()
 
         // Add active task
@@ -259,7 +297,8 @@ import org.junit.runner.RunWith
         onView(withText(TITLE1)).check(matches(not(isDisplayed())))
     }
 
-    @Test fun createTwoTasks_deleteOneTask() {
+    @Test
+    fun createTwoTasks_deleteOneTask() {
         // Add 2 active tasks
         createTask(TITLE1, DESCRIPTION)
         createTask(TITLE2, DESCRIPTION)
@@ -276,7 +315,8 @@ import org.junit.runner.RunWith
         onView(withText(TITLE2)).check(doesNotExist())
     }
 
-    @Test fun markTaskAsCompleteOnDetailScreen_taskIsCompleteInList() {
+    @Test
+    fun markTaskAsCompleteOnDetailScreen_taskIsCompleteInList() {
         viewAllTasks()
 
         // Add 1 active task
@@ -292,11 +332,12 @@ import org.junit.runner.RunWith
         onView(withContentDescription(toolbarNavigationContentDescription)).perform(click())
 
         // Check that the task is marked as completed
-        onView(allOf(withId(R.id.complete),
-                hasSibling(withText(TITLE1)))).check(matches(isChecked()))
+        onView(allOf(withId(R.id.complete), hasSibling(withText(TITLE1))))
+            .check(matches(isChecked()))
     }
 
-    @Test fun markTaskAsActiveOnDetailScreen_taskIsActiveInList() {
+    @Test
+    fun markTaskAsActiveOnDetailScreen_taskIsActiveInList() {
         viewAllTasks()
 
         // Add 1 completed task
@@ -313,11 +354,12 @@ import org.junit.runner.RunWith
         onView(withContentDescription(toolbarNavigationContentDescription)).perform(click())
 
         // Check that the task is marked as active
-        onView(allOf(withId(R.id.complete),
-                hasSibling(withText(TITLE1)))).check(matches(not(isChecked())))
+        onView(allOf(withId(R.id.complete), hasSibling(withText(TITLE1))))
+            .check(matches(not(isChecked())))
     }
 
-    @Test fun markTaskAsAcompleteAndActiveOnDetailScreen_taskIsActiveInList() {
+    @Test
+    fun markTaskAsAcompleteAndActiveOnDetailScreen_taskIsActiveInList() {
         viewAllTasks()
 
         // Add 1 active task
@@ -336,11 +378,12 @@ import org.junit.runner.RunWith
         onView(withContentDescription(toolbarNavigationContentDescription)).perform(click())
 
         // Check that the task is marked as active
-        onView(allOf(withId(R.id.complete),
-                hasSibling(withText(TITLE1)))).check(matches(not(isChecked())))
+        onView(allOf(withId(R.id.complete), hasSibling(withText(TITLE1))))
+            .check(matches(not(isChecked())))
     }
 
-    @Test fun markTaskAsActiveAndCompleteOnDetailScreen_taskIsCompleteInList() {
+    @Test
+    fun markTaskAsActiveAndCompleteOnDetailScreen_taskIsCompleteInList() {
         viewAllTasks()
 
         // Add 1 completed task
@@ -360,11 +403,12 @@ import org.junit.runner.RunWith
         onView(withContentDescription(toolbarNavigationContentDescription)).perform(click())
 
         // Check that the task is marked as active
-        onView(allOf(withId(R.id.complete),
-                hasSibling(withText(TITLE1)))).check(matches(isChecked()))
+        onView(allOf(withId(R.id.complete), hasSibling(withText(TITLE1))))
+            .check(matches(isChecked()))
     }
 
-    @Test fun orientationChange_FilterActivePersists() {
+    @Test
+    fun orientationChange_FilterActivePersists() {
 
         // Add a completed task
         createTask(TITLE1, DESCRIPTION)
@@ -383,7 +427,8 @@ import org.junit.runner.RunWith
         onView(withText(TITLE1)).check(doesNotExist())
     }
 
-    @Test fun orientationChange_FilterCompletedPersists() {
+    @Test
+    fun orientationChange_FilterCompletedPersists() {
 
         // Add a completed task
         createTask(TITLE1, DESCRIPTION)
@@ -400,11 +445,13 @@ import org.junit.runner.RunWith
 
         // then nothing changes
         onView(withText(TITLE1)).check(matches(isDisplayed()))
-        onView(withText(R.string.label_completed)).check(matches(isDisplayed()))
+        onView(withText(string.label_completed)).check(matches(isDisplayed()))
     }
 
     // Blinking cursor after rotation breaks this in API 19
-    @Test @SdkSuppress(minSdkVersion = 21) fun orientationChange_DuringEdit_ChangePersists() {
+    @Test
+    @SdkSuppress(minSdkVersion = 21)
+    fun orientationChange_DuringEdit_ChangePersists() {
         // Add a completed task
         createTask(TITLE1, DESCRIPTION)
 
@@ -416,7 +463,7 @@ import org.junit.runner.RunWith
 
         // Change task title (but don't save)
         onView(withId(R.id.add_task_title))
-                .perform(replaceText(TITLE2), closeSoftKeyboard()) // Type new task title
+            .perform(replaceText(TITLE2), closeSoftKeyboard()) // Type new task title
 
         // Rotate the screen
         currentActivity.rotateOrientation()
@@ -426,7 +473,9 @@ import org.junit.runner.RunWith
     }
 
     // Blinking cursor after rotation breaks this in API 19
-    @Test @SdkSuppress(minSdkVersion = 21) fun orientationChange_DuringEdit_NoDuplicate() {
+    @Test
+    @SdkSuppress(minSdkVersion = 21)
+    fun orientationChange_DuringEdit_NoDuplicate() {
         // Add a completed task
         createTask(TITLE1, DESCRIPTION)
 
@@ -441,9 +490,11 @@ import org.junit.runner.RunWith
 
         // Edit task title and description
         onView(withId(R.id.add_task_title))
-                .perform(replaceText(TITLE2), closeSoftKeyboard()) // Type new task title
-        onView(withId(R.id.add_task_description)).perform(replaceText(DESCRIPTION),
-                closeSoftKeyboard()) // Type new task description and close the keyboard
+            .perform(replaceText(TITLE2), closeSoftKeyboard()) // Type new task title
+        onView(withId(R.id.add_task_description)).perform(
+            replaceText(DESCRIPTION),
+            closeSoftKeyboard()
+        ) // Type new task description and close the keyboard
 
         // Save the task
         onView(withId(R.id.fab_edit_task_done)).perform(click())
@@ -455,7 +506,8 @@ import org.junit.runner.RunWith
         onView(withItemText(TITLE1)).check(doesNotExist())
     }
 
-    @Test fun noTasks_AllTasksFilter_AddTaskViewVisible() {
+    @Test
+    fun noTasks_AllTasksFilter_AddTaskViewVisible() {
         // Given an empty list of tasks, make sure "All tasks" filter is on
         viewAllTasks()
 
@@ -463,7 +515,8 @@ import org.junit.runner.RunWith
         onView(withId(R.id.noTasksAdd)).check(matches(isDisplayed()))
     }
 
-    @Test fun noTasks_CompletedTasksFilter_AddTaskViewNotVisible() {
+    @Test
+    fun noTasks_CompletedTasksFilter_AddTaskViewNotVisible() {
         // Given an empty list of tasks, make sure "All tasks" filter is on
         viewCompletedTasks()
 
@@ -471,7 +524,8 @@ import org.junit.runner.RunWith
         onView(withId(R.id.noTasksAdd)).check(matches(not(isDisplayed())))
     }
 
-    @Test fun noTasks_ActiveTasksFilter_AddTaskViewNotVisible() {
+    @Test
+    fun noTasks_ActiveTasksFilter_AddTaskViewNotVisible() {
         // Given an empty list of tasks, make sure "All tasks" filter is on
         viewActiveTasks()
 
@@ -481,17 +535,17 @@ import org.junit.runner.RunWith
 
     private fun viewAllTasks() {
         onView(withId(R.id.menu_filter)).perform(click())
-        onView(withText(R.string.nav_all)).perform(click())
+        onView(withText(string.nav_all)).perform(click())
     }
 
     private fun viewActiveTasks() {
         onView(withId(R.id.menu_filter)).perform(click())
-        onView(withText(R.string.nav_active)).perform(click())
+        onView(withText(string.nav_active)).perform(click())
     }
 
     private fun viewCompletedTasks() {
         onView(withId(R.id.menu_filter)).perform(click())
-        onView(withText(R.string.nav_completed)).perform(click())
+        onView(withText(string.nav_completed)).perform(click())
     }
 
     private fun createTask(title: String, description: String) {
@@ -499,10 +553,14 @@ import org.junit.runner.RunWith
         onView(withId(R.id.fab_add_task)).perform(click())
 
         // Add task title and description
-        onView(withId(R.id.add_task_title)).perform(typeText(title),
-                closeSoftKeyboard()) // Type new task title
-        onView(withId(R.id.add_task_description)).perform(typeText(description),
-                closeSoftKeyboard()) // Type new task description and close the keyboard
+        onView(withId(R.id.add_task_title)).perform(
+            typeText(title),
+            closeSoftKeyboard()
+        ) // Type new task title
+        onView(withId(R.id.add_task_description)).perform(
+            typeText(description),
+            closeSoftKeyboard()
+        ) // Type new task description and close the keyboard
 
         // Save the task
         onView(withId(R.id.fab_edit_task_done)).perform(click())
@@ -513,5 +571,5 @@ import org.junit.runner.RunWith
     }
 
     private fun getText(stringId: Int) =
-            tasksActivityTestRule.activity.resources.getString(stringId)
+        tasksActivityTestRule.activity.resources.getString(stringId)
 }
