@@ -16,9 +16,9 @@
 package com.example.android.architecture.blueprints.todoapp
 
 import android.content.Context
-import android.support.v4.widget.SwipeRefreshLayout
 import android.util.AttributeSet
 import android.view.View
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 
 /**
  * Extends [SwipeRefreshLayout] to support non-direct descendant scrolling views.
