@@ -1,74 +1,78 @@
-/*
- *  Copyright 2017 Google Inc.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+/* Looking for the SingleLiveEvent class? We are now using the Event class. Read more:
+https://medium.com/androiddevelopers/livedata-with-snackbar-navigation-and-other-events-the-singleliveevent-case-ac2622673150
  */
-package com.example.android.architecture.blueprints.todoapp
-
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.Observer
-import android.support.annotation.MainThread
-import android.util.Log
-
-import java.util.concurrent.atomic.AtomicBoolean
-
-/**
- * A lifecycle-aware observable that sends only new updates after subscription, used for events like
- * navigation and Snackbar messages.
- *
- *
- * This avoids a common problem with events: on configuration change (like rotation) an update
- * can be emitted if the observer is active. This LiveData only calls the observable if there's an
- * explicit call to setValue() or call().
- *
- *
- * Note that only one observer is going to be notified of changes.
- */
-class SingleLiveEvent<T> : MutableLiveData<T>() {
-
-    private val pending = AtomicBoolean(false)
-
-    @MainThread
-    override fun observe(owner: LifecycleOwner, observer: Observer<T>) {
-
-        if (hasActiveObservers()) {
-            Log.w(TAG, "Multiple observers registered but only one will be notified of changes.")
-        }
-
-        // Observe the internal MutableLiveData
-        super.observe(owner, Observer<T> { t ->
-            if (pending.compareAndSet(true, false)) {
-                observer.onChanged(t)
-            }
-        })
-    }
-
-    @MainThread
-    override fun setValue(t: T?) {
-        pending.set(true)
-        super.setValue(t)
-    }
-
-    /**
-     * Used for cases where T is Void, to make calls cleaner.
-     */
-    @MainThread
-    fun call() {
-        value = null
-    }
-
-    companion object {
-        private const val TAG = "SingleLiveEvent"
-    }
-}
+//
+///*
+// *  Copyright 2017 Google Inc.
+// *
+// *  Licensed under the Apache License, Version 2.0 (the "License");
+// *  you may not use this file except in compliance with the License.
+// *  You may obtain a copy of the License at
+// *
+// *      http://www.apache.org/licenses/LICENSE-2.0
+// *
+// *  Unless required by applicable law or agreed to in writing, software
+// *  distributed under the License is distributed on an "AS IS" BASIS,
+// *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// *  See the License for the specific language governing permissions and
+// *  limitations under the License.
+// */
+//package com.example.android.architecture.blueprints.todoapp
+//
+//import androidx.lifecycle.LifecycleOwner
+//import androidx.lifecycle.MutableLiveData
+//import androidx.lifecycle.Observer
+//import androidx.annotation.MainThread
+//import android.util.Log
+//
+//import java.util.concurrent.atomic.AtomicBoolean
+//
+///**
+// * A lifecycle-aware observable that sends only new updates after subscription, used for events like
+// * navigation and Snackbar messages.
+// *
+// *
+// * This avoids a common problem with events: on configuration change (like rotation) an update
+// * can be emitted if the observer is active. This LiveData only calls the observable if there's an
+// * explicit call to setValue() or call().
+// *
+// *
+// * Note that only one observer is going to be notified of changes.
+// */
+//class SingleLiveEvent<T> : MutableLiveData<T>() {
+//
+//    private val pending = AtomicBoolean(false)
+//
+//    @MainThread
+//    override fun observe(owner: LifecycleOwner, observer: Observer<T>) {
+//
+//        if (hasActiveObservers()) {
+//            Log.w(TAG, "Multiple observers registered but only one will be notified of changes.")
+//        }
+//
+//        // Observe the internal MutableLiveData
+//        super.observe(owner, Observer<T> { t ->
+//            if (pending.compareAndSet(true, false)) {
+//                observer.onChanged(t)
+//            }
+//        })
+//    }
+//
+//    @MainThread
+//    override fun setValue(t: T?) {
+//        pending.set(true)
+//        super.setValue(t)
+//    }
+//
+//    /**
+//     * Used for cases where T is Void, to make calls cleaner.
+//     */
+//    @MainThread
+//    fun call() {
+//        value = null
+//    }
+//
+//    companion object {
+//        private const val TAG = "SingleLiveEvent"
+//    }
+//}
